+++ src/l1j/jrwz/server/model/map/L1WorldMap.java
@@ -67,7 +67,7 @@ public class L1WorldMap {
      *            マップID
      * @return マップ情報を保持する、L1Mapオブジェクト。
      */
-    public L1Map getMap(short mapId) {
+    public L1Map getMap(int mapId) {
         L1Map map = _maps.get((int) mapId);
         if (map == null) { // マップ情報が無い
             map = L1Map.newNull(); // 何もしないMapを返す。
